Move methods from TfsTeamSettingsCore to TfsTeamSettingsTool
